make aggregate_regions_vectorised() generic, aggregating any matrix
Instead of passing U, V, etc., make this function generic. It accepts a
function, aggregates it, and returns the aggregated matrix.

This function therefore has one new argument (the matrix) and one fewer
(the "inplace" flag). This function can potentially be employed as part
of a broader function to aggregate a whole supply and use inventory.

 On branch constructs
 Changes to be committed:
	modified:   pySUT.py
	modified:   pySUT_UnitTest.py